Brought the branch up to date with dev (155 commits: supply-plausibility
guard, declared reading channel, weekly reports, gateway/security work).

Conflicts were all in telemetry-service and resolved keeping both sides:

- BfmReadingService: dev's supplyPlausibilityGuard collaborator and the
  branch's ocrProviderResolver/ocrProviderRegistry now sit side by side.
  The OCR dispatch (resolve per tenant, then FlowVision or the registered
  provider) is unchanged, and it is still the only OCR entry point.
- BfmReadingService tests that dev touched had to pass the two new
  constructor arguments; they pass null, which keeps them on the built-in
  FlowVision path exactly as before.

Tenants with no ocr_* config still resolve to null settings and take the
original FlowVision path, so nothing existing changes behaviour.

mvn test on telemetry-service: 1662 tests, 0 failures.

+++ backend/analytics-service/src/main/java/org/arghyam/jalsoochak/analytics/dto/event/MeterReadingEvent.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -13,8 +15,14 @@ public class MeterReadingEvent {
     private Integer tenantId;
     private Integer schemeId;
     private Integer userId;
-    private Integer extractedReading;
-    private Integer confirmedReading;
+    /**
+     * Readings are {@code BigDecimal} because the meters carry a decimal digit and
+     * {@code flow_reading_table} records it. An event published by a telemetry-service still on the
+     * old contract carries a whole number, which deserialises into this field unchanged — so this
+     * service can be deployed first, and must be, since it owns the column widening.
+     */
+    private BigDecimal extractedReading;
+    private BigDecimal confirmedReading;
     private Integer confidence;
     private String imageUrl;
     private String readingAt;
