+++ backend/telemetry-service/src/main/java/org/arghyam/jalsoochak/telemetry/dto/event/WaterQuantityEvent.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -15,7 +17,12 @@ public class WaterQuantityEvent {
     private Integer tenantId;
     private Integer schemeId;
     private Integer userId;
-    private Integer waterQuantity;
+    /**
+     * The correction paths' daily volume in the meter's native m&sup3;, carried at the precision they
+     * computed it — the difference of two {@code NUMERIC} readings. Analytics converts to the litres its
+     * column stores; this service does not, so that the two never disagree about the unit.
+     */
+    private BigDecimal waterQuantity;
     private Integer submissionStatus;
     private String outageReason;
     private String nonSubmissionReason;
